get_form_fast_api/requirements.txt
fastapi==0.100.0           
uvicorn==0.23.1            
pydantic==2.10.2          
python-dotenv==1.0.0 
pymongo==4.10.1